% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-utils.R
\name{generate_gryn_sites}
\alias{generate_gryn_sites}
\title{Save the GYRN historical and current WQ sites as a list}
\usage{
generate_gryn_sites(
  sites = c("11NPSWRD_WQX-GRTE_SNR01", "11NPSWRD_WQX-GRTE_SNR02",
    "11NPSWRD_WQX-YELL_MDR", "11NPSWRD_WQX-YELL_LM000.5M", "11NPSWRD_WQX-YELL_SB015.7A",
    "11NPSWRD_WQX-YELL_YS549.7M", "11NPSWRD_WQX-BICA_BHR2", "11NPSWRD_WQX-BICA_BHR1",
    "11NPSWRD_WQX-BICA_SHR1"),
  legacy_project = c("YELLWQ01_LEGACY"),
  legacy_sites = c("11NPSWRD_WQX-YELL_MD133.2T", "11NPSWRD_WQX-YELL_LM000.5M",
    "11NPSWRD_WQX-YELL_SB015.7A", "11NPSWRD_WQX-YELL_YS549.7M")
)
}
\arguments{
\item{sites}{Optional. A vector of WQP of site IDs,
e.g. c("11NPSWRD_WQX-GRTE_SNR01", "11NPSWRD_WQX-GRTE_SNR02")}

\item{legacy_project}{Optional. A vector of legacy projects that legacy_sites are stored under.}

\item{legacy_sites}{Optional. A vector of legacy sites that are stored under legacy_project.}
}
\value{
List with names sites, legacy project, and legacy sites.
}
\description{
Save the GYRN historical and current WQ sites as a list
}
\examples{
generate_gryn_sites()
generate_gryn_sites(sites = c("11NPSWRD_WQX-GRTE_SNR01", "11NPSWRD_WQX-GRTE_SNR02"),
 legacy_project = NA, legacy_sites = NA)
}
