% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-utils.R
\name{grab_gryn_data}
\alias{grab_gryn_data}
\title{Pull all GRYN sites into a list}
\usage{
grab_gryn_data()
}
\value{
a list of GRYN monitored WQP data frames
}
\description{
Pull all GRYN sites into a list
}
\examples{
gryn_site_dfs <- grab_gryn_data()
}
