% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-utils.R
\name{grab_wq_data}
\alias{grab_wq_data}
\title{Pull Site Data from the WQ Portal}
\usage{
grab_wq_data(site_list, data_profile = "resultPhysChem", progress_bar = TRUE)
}
\arguments{
\item{site_list}{A list of WQP site IDs}

\item{data_profile}{The type of data to pull. Default is physical/chemical.}

\item{progress_bar}{T/F. Do you want to display a progress bar? Helpful if you
expect the pull to take a long time. Default = T.}
}
\value{
A list of WQP data frames for input site IDs
}
\description{
Pull Site Data from the WQ Portal
}
\examples{
site_name_list <- as.list(c("11NPSWRD_WQX-GRTE_SNR01", "11NPSWRD_WQX-GRTE_SNR02"))
site_df_list <- grab_wq_data(site_name_list)
}
