% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-utils.R
\name{remove_fields}
\alias{remove_fields}
\title{Selects relevant fields from WQP pull}
\usage{
remove_fields(site)
}
\arguments{
\item{site}{a WQP resultPhysChem data frame}
}
\value{
A data frame with only relevant columns remaining
}
\description{
Selects relevant fields from WQP pull
}
\examples{
MDR1 <- dataRetrieval::readWQPdata(siteid = "11NPSWRD_WQX-YELL_MDR",
dataProfile = "resultPhysChem")
MDR1 <- remove_fields(MDR1)
}
