% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-utils.R
\name{remove_fields_site_dfs}
\alias{remove_fields_site_dfs}
\title{Selects relevant fields from each data frame in WQP pull list}
\usage{
remove_fields_site_dfs(site_dfs)
}
\arguments{
\item{site_dfs}{the list of data frames pulled from WQP}
}
\value{
updated list of site data frames
}
\description{
Selects relevant fields from each data frame in WQP pull list
}
\examples{
site_list <- as.list(c("11NPSWRD_WQX-GRTE_SNR01", "11NPSWRD_WQX-GRTE_SNR02"))
site_dfs <- grab_wq_data(site_list)
site_dfs <- remove_fields_site_dfs(site_dfs)
}
