% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-utils.R
\name{remove_fields_site_list}
\alias{remove_fields_site_list}
\title{Selects relevant fields from each data frame in WQP pull list}
\usage{
remove_fields_site_list(site_list)
}
\arguments{
\item{site_list}{the list of data frames pulled from WQP}
}
\value{
updated site_list of data frames
}
\description{
Selects relevant fields from each data frame in WQP pull list
}
\examples{
site_list <- grab_gryn_data()
site_list <- remove_fields_site_list(site_list)
}
