% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-utils.R
\name{generate_gryn_sites}
\alias{generate_gryn_sites}
\title{Saves the GYRN historical and current WQ sites as a list.}
\usage{
generate_gryn_sites(
  sites = c("11NPSWRD_WQX-GRTE_SNR01", "11NPSWRD_WQX-GRTE_SNR02",
    "11NPSWRD_WQX-YELL_MDR", "11NPSWRD_WQX-YELL_MD133.2T", "11NPSWRD_WQX-YELL_LM000.5M",
    "11NPSWRD_WQX-YELL_SB015.7A", "11NPSWRD_WQX-YELL_YS549.7M", "11NPSWRD_WQX-BICA_BHR2",
    "11NPSWRD_WQX-BICA_BHR1", "11NPSWRD_WQX-BICA_SHR1")
)
}
\arguments{
\item{sites}{A vector of WQP of site IDs,
e.g. c("11NPSWRD_WQX-GRTE_SNR01", "11NPSWRD_WQX-GRTE_SNR02"). Defaults to GRYN WQ sites.}
}
\value{
A list of the site names, where each element of the list is a site name.
}
\description{
Saves the GYRN historical and current WQ sites as a list.
}
\examples{
generate_gryn_sites()
generate_gryn_sites(sites = c("11NPSWRD_WQX-GRTE_SNR01", "11NPSWRD_WQX-GRTE_SNR02"))
}
